Break prefork out of cost_basis.py to fix circular dependency
